UeCherab: Split d3d into submodule
Going to add bolometer and later other diagnostics, so
split into separate files.
